ISSUE-263: Moved some classes around
